Reorganise files into components part II
